# shared/openapi.yaml
openapi: "3.1.0"
info:
  title: Matter-rs Game Server
  version: "0.1.0"
  description: Multiplayer game server API
paths:
...
  /auth/me:
    get:
      summary: Get the current signed-in player
      responses:
        "200":
          description: Current user
        "401":
          description: No active session
...